requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cpu 
torchvision==0.11.2+cpu
Flask==1.0.3
numpy==1.22.1
Pillow==8.1.1
gunicorn==19.9.0